requirements.txt
flask
opencv-python
numpy
onnxruntime
google-generativeai
python-dotenv
